+++ RingPublishingGDPR/src/main/java/com/ringpublishing/gdpr/RingPublishingGDPRActivity.java
@@ -99,6 +99,7 @@ public class RingPublishingGDPRActivity extends AppCompatActivity implements Rin
     @Override
     public void hide(FormView formView)
     {
+        RingPublishingGDPR.getInstance().notifyConsentsUpdated();
         setResult(RESULT_OK);
         finish();
     }