+++ demo/src/main/java/com/ringpublishing/gdpr/demo/DemoApplication.java
@@ -57,6 +57,24 @@ public class DemoApplication extends MultiDexApplication
         // When (in you usage context) GDPR does not apply, use initialize method with extra parameter gdprApplies set to false
         // ringPublishingGDPR.initialize(false, this, appTenantId, appBrandingName, ringPublishingGDPRUIConfig);
 
+        // Optional listener that informs application about saving or updating consents.
+        // You can add more listeners on each place where you initialize SDK.
+        // Remember call ringPublishingGDPR.removeRingPublishingGDPRListeners() on your object destroy
+        ringPublishingGDPR.addRingPublishingGDPRListeners(() ->
+        {
+            // Here you can check if user agreed on all available vendors
+            if(RingPublishingGDPR.getInstance().areVendorConsentsGiven())
+            {
+                // If you app uses SDK's from vendors, which are not part of the official TCF 2.0 vendor list
+                // you can use this flag to check if you can enable / initialize them as user agreed on all vendors
+                initializeApplicationExternalLibraries();
+            }
+        });
     }
 
+    private void initializeApplicationExternalLibraries()
+    {
+        // Here we can initialize or update initialization all application libraries
+        Toast.makeText(DemoApplication.this, R.string.toast_libraries_application, Toast.LENGTH_SHORT).show();
+    }
 }