+++ demo/src/main/java/com/ringpublishing/gdpr/demo/SplashActivity.java
@@ -3,6 +3,7 @@ package com.ringpublishing.gdpr.demo;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
+import android.widget.Toast;
 
 import com.ringpublishing.gdpr.RingPublishingGDPR;
 import com.ringpublishing.gdpr.demo.databinding.ActivitySplashBinding;
@@ -53,31 +54,34 @@ public class SplashActivity extends AppCompatActivity
 
     private void showAppContent()
     {
-        // Here you can check if user agreed for all available vendors
+        // Here you can check if user agreed on all available vendors
         if(RingPublishingGDPR.getInstance().areVendorConsentsGiven())
         {
-            initializeExternalLibraries();
+            // If you app uses SDK's from vendors, which are not part of the official TCF 2.0 vendor list
+            // you can use this flag to check if you can enable / initialize them as user agreed on all vendors
+            initializeSplashExternalLibraries();
         }
         startActivity(new Intent(this, MainActivity.class));
         finish();
-        // If you app uses SDK's from vendors, which are not part of the official TCF 2.0 vendor list
-        // you can use this flag to check if you can enable / initialize them as user agreed for all vendors
+
     }
 
-    private void initializeExternalLibraries()
+    private void initializeSplashExternalLibraries()
     {
-        // Here you can initialize external libraries, because consent screen was already visited by user.
+        // Here you can initialize external libraries initialized in your Splash screen,
+        // becaue consent screen has been already presented to user.
+        Toast.makeText(this, R.string.toast_libraries_splash, Toast.LENGTH_SHORT).show();
     }
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data)
     {
+        super.onActivityResult(requestCode, resultCode, data);
         if (requestCode == REQUEST_CODE_OPEN_CONSENT)
         {
             Log.i(TAG, "SplashActivity onActivityResult showAppContent");
             showAppContent();
         }
-        super.onActivityResult(requestCode, resultCode, data);
     }
 
 }