+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 
 ### Features
 
-* Public interface has been extended about one additional properties "publicConsents" which should be used by EventsSDK
+* New value 'RingPublishing_PublicConsents' stored in user defaults (should be used for other purposes than Ad Server (public))
 * Public interface has been extended about two additional properties  "areVendorConsentsGiven" which can be used by other Ring Publishing modules, e.g. Ad Server.
+* Remove from demo AndroidMainfest.xml  <activity android:name="com.ringpublishing.gdpr.RingPublishingGDPRActivity" android:theme="@style/GDPRStyle" /> move it to RingPublishingGDPR
 
 1.1.0 Release notes (2020-11-13)
 =============================================================